requirements.txt
flask
torch
torchvision
timm
pillow
gunicorn